feat: adjusted the database to take integers in the prediction and predictor DAOs

+++ Apromore-Database/src/main/java/org/apromore/dao/PredictorRepository.java
@@ -24,6 +24,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PredictorRepository extends JpaRepository<PredictorDao, Long> {
+public interface PredictorRepository extends JpaRepository<PredictorDao, Integer> {
     List<PredictorDao> findAllByLogId(int logId);
 }
